+++ .gitignore
@@ -2,3 +2,4 @@
 node_modules/
 config.json
 npm-debug.log
+build/
