+++ .travis.yml
@@ -1,0 +1,13 @@
+language: node_js
+node_js:
+  - "6"
+  - "5"
+  - "4"
+sudo: required
+dist: trusty
+env:
+  - EVENTSTORE_HOST="127.0.0.1" ES_DOWNLOAD_URL="http://download.geteventstore.com/binaries/EventStore-OSS-Ubuntu-14.04-v3.9.1.tar.gz"
+install:
+  - npm install
+script:
+  - ./setup.sh
